fix bug where integers would be parsed as i32 instead of being i64 and then cast to the smaller size as needed

+++ src/parse/mod.rs
@@ -1916,7 +1916,7 @@ impl Parser {
             TokenKind::Integer(n) => {
                 self.advance()?;
                 Ok(Expr {
-                    kind: ExprKind::Constant(Const::Int(n)),
+                    kind: ExprKind::Constant(Const::Int(n as i32)),
                     start,
                     end,
                     source,
